// src/models/professional-experience.ts
export class ProfessionalExperienceModel {
  company: string = '';
  initialMonth: string = '';
  initialYear: number | null = null;
  endMonth: string = '';
  endYear: number | null = null;
  position: string = '';
  country: string = '';
  workModel: string = '';
  actualWork: boolean = false;
  description: string = '';
}